Sprint#6 Task#291 Startuje zdrapywacze asynchronicznie i zwraca wiadomość o rozpoczęciu pracy
- Startowanie metody odpalającej zdrapywacze w osobnym wątku
- refaktoryzacja kodu: przeniesienie logiki z RobotScrappingStarted do ScrappingService
